+++ Marlin/src/module/ft_motion/trajectory_poly6.cpp
@@ -24,7 +24,7 @@
 
 #if ENABLED(FT_MOTION)
 
-#include "poly6_trajectory_generator.h"
+#include "trajectory_poly6.h"
 #include <math.h>
 #include "../ft_motion.h"
