+++ Marlin/src/module/ft_motion/trajectory_poly6.h
